Refactors trimmers so they only implement valid()
Before the trimmers would all have code to create an iterator, work through it
and output results. This commit moves all this code into a base class which
provides a protected `bool valid(MatrixPtr)` method for trimmers to override.

There was a small problem with dealing with different types of matrices - which
is overcome by creating a base Trimmer class which does not depend on the matrix
at all, and a __Trimmer<Matrix> class which does. This way TrimmerBuilder still
uses Trimmer and does not care about the type of matrices while the trimmers
themselves can specify whether to use __Trimmer<IntMatrix> or
__Trimmer<EquivQuiverMatrix>.

None of the virtual trimmers had virtual destructors - so these were added too.